added react-router-bootstrap and reactjs-popup to package.json and it also automatically updated package-lock.json. moved svg images into a folder and corrected references to them. deleted service workers and react logo
