+++ garden-harvest/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import Routes from "./Routes";
+import Routes from "./components/Routes";
 
 function App() {
   return (
